+++ src/tedi/components/form/date-calendar/components/date-calendar-month-grid/date-calendar-month-grid.tsx
@@ -1,6 +1,6 @@
 import { useLabels } from '../../../../../providers/label-provider';
 import { Text } from '../../../../base/typography/text/text';
-import { PickerGrid } from '../../date-field-grid';
+import { PickerGrid } from '../../date-calendar-grid';
 
 export interface MonthGridProps {
   /*