+++ src/tedi/components/form/date-calendar/components/date-calendar-year-grid/date-calendar-year-grid.tsx
@@ -1,5 +1,5 @@
 import { useLabels } from '../../../../../providers/label-provider';
-import { PickerGrid } from '../../date-field-grid';
+import { PickerGrid } from '../../date-calendar-grid';
 
 export interface YearGridProps {
   /*