+++ CHANGELOG.md
@@ -1,3 +1,24 @@
+# [16.1.0-rc.3](https://github.com/TEDI-Design-System/react/compare/react-16.1.0-rc.2...react-16.1.0-rc.3) (2026-03-11)
+
+
+### Features
+
+* **dropdown:** TEDI-Ready Dropdown component [#94](https://github.com/TEDI-Design-System/react/issues/94) ([#549](https://github.com/TEDI-Design-System/react/issues/549)) ([836039f](https://github.com/TEDI-Design-System/react/commit/836039ff317b32fed2dc0c47a8b495ddcffe6e91))
+
+# [16.1.0-rc.2](https://github.com/TEDI-Design-System/react/compare/react-16.1.0-rc.1...react-16.1.0-rc.2) (2026-03-11)
+
+
+### Bug Fixes
+
+* **list:** fix mobile padding [#552](https://github.com/TEDI-Design-System/react/issues/552) ([#553](https://github.com/TEDI-Design-System/react/issues/553)) ([2e93804](https://github.com/TEDI-Design-System/react/commit/2e93804f5b1bf7662eef9ce3517c36fa637f61e9))
+
+# [16.1.0-rc.1](https://github.com/TEDI-Design-System/react/compare/react-16.0.1-rc.2...react-16.1.0-rc.1) (2026-03-10)
+
+
+### Features
+
+* **theme-provider:** theme provider changes [#499](https://github.com/TEDI-Design-System/react/issues/499) ([#550](https://github.com/TEDI-Design-System/react/issues/550)) ([07fa36c](https://github.com/TEDI-Design-System/react/commit/07fa36c1a26d5f9b20b6cada92b1ab92e098f587))
+
 ## [16.0.1-rc.2](https://github.com/TEDI-Design-System/react/compare/react-16.0.1-rc.1...react-16.0.1-rc.2) (2026-03-05)